get note - created at as long type

+++ app/src/main/java/com/apps/home/notewidget/utils/DatabaseHelper.java
@@ -489,7 +489,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
                     Note note = new Note();
                     note.setId(cursor.getInt(cursor.getColumnIndexOrThrow(Constants.ID_COL)));
-                    note.setCreatedAt(cursor.getInt(cursor.getColumnIndexOrThrow(Constants.MILLIS_COL)));
+                    note.setCreatedAt(cursor.getLong(cursor.getColumnIndexOrThrow(Constants.MILLIS_COL)));
                     note.setTitle(cursor.getString(cursor.getColumnIndexOrThrow(Constants.NOTE_TITLE_COL)));
                     note.setNote(cursor.getString(cursor.getColumnIndexOrThrow(Constants.NOTE_TEXT_COL)));
                     note.setFolderId(cursor.getInt(cursor.getColumnIndexOrThrow(Constants.FOLDER_ID_COL)));
